Replace old codes.interface module
