+++ bluemira/codes/interface.py
@@ -18,155 +18,72 @@
 #
 # You should have received a copy of the GNU Lesser General Public
 # License along with bluemira; if not, see <https://www.gnu.org/licenses/>.
+"""Base classes for solvers using external codes."""
 
-"""
-The bluemira external code wrapper
-"""
-from __future__ import annotations
-
-import string
-import subprocess  # noqa :S404
-from abc import ABC, abstractmethod
-from enum import Enum
-from typing import Callable, Dict, List, Optional, Union
+import abc
+from typing import Any, Callable, Dict, List, Optional, Union
 
 from bluemira.base.constants import raw_uc
 from bluemira.base.look_and_feel import bluemira_warn
 from bluemira.base.parameter import ParameterFrame
+from bluemira.base.solver import SolverABC, Task
 from bluemira.codes.error import CodesError
-from bluemira.codes.utilities import (
-    LogPipe,
-    add_mapping,
-    get_recv_mapping,
-    get_send_mapping,
-)
-
-__all__ = [
-    "FileProgramInterface",
-    "RunMode",
-    "Setup",
-    "Run",
-    "Teardown",
-]  # "ApplicationProgramInterface"]
+from bluemira.codes.utilities import get_recv_mapping, get_send_mapping, run_subprocess
 
 
-class RunMode(Enum):
-    """Defines the possible runmode"""
+class CodesTask(Task):
+    """
+    Base class for a task used by a solver for an external code.
+    """
 
-    def __call__(self, obj, *args, **kwargs):
+    def __init__(self, params: ParameterFrame, codes_name: str) -> None:
+        super().__init__(params)
+        self._name = codes_name
+
+    def _run_subprocess(self, command: List[str], **kwargs):
         """
-        Call function of object with lowercase name of
-        enum
+        Run a subprocess command and raise a CodesError if it returns a
+        non-zero exit code.
+        """
+        return_code = run_subprocess(command, **kwargs)
+        if return_code != 0:
+            raise CodesError(
+                f"'{self._name}' subprocess task exited with non-zero error code "
+                f"'{return_code}'."
+            )
+
+
+class CodesSetup(CodesTask):
+    """
+    Base class for setup tasks of a solver for an external code.
+    """
+
+    def _get_new_inputs(
+        self, remapper: Optional[Union[Callable, Dict[str, str]]] = None
+    ) -> Dict[str, float]:
+        """
+        Retrieve inputs values to the external code from this tasks'
+        ParameterFrame.
+
+        Convert the inputs' units to those used by the external code.
 
         Parameters
         ----------
-        obj: instance
-            instance of class the function will come from. If object is not specified,
-            self will be used.
-        *args
-           args of function
-        **kwargs
-           kwargs of function
-        Returns
-        -------
-        function result
-        """
-        name = f"_{self.name.lower()}"
-        if hasattr(obj, name):
-            func = getattr(obj, name)
-            return func(*args, **kwargs)
-
-
-class Task:
-    """
-    Task integration
-
-    Parameters
-    ----------
-    parent: obj
-        parent file interface object
-
-    """
-
-    def __init__(self, parent):
-        self.parent = parent
-
-    def _run_subprocess(self, command, **kwargs):
-        stdout = LogPipe("print")
-        stderr = LogPipe("error")
-
-        kwargs["cwd"] = kwargs.get("cwd", self.parent.run_dir)
-        kwargs.pop("shell", None)  # Protect against user input
-
-        with subprocess.Popen(  # noqa :S603
-            command, stdout=stdout, stderr=stderr, shell=False, **kwargs  # noqa :S603
-        ) as s:
-            stdout.close()
-            stderr.close()
-
-        if s.returncode:
-            raise CodesError(f"{self.parent.NAME} exited with a non zero exit code")
-
-
-class Setup(Task):
-    """
-    Generic Setup Task
-
-    Parameters
-    ----------
-    parent: obj
-        parent file interface object
-    params: ParameterFrame
-        ParameterFrame for interface
-
-    """
-
-    def __init__(self, parent, *args, params=None, problem_settings=None, **kwargs):
-        super().__init__(parent)
-
-        self._problem_settings = problem_settings if problem_settings is not None else {}
-        self.set_parameters(params)
-
-    def set_parameters(self, params):
-        """
-        Set parameter mappings and add parameters to interface
-
-        Parameters
-        ----------
-        params: ParameterFrame
-
-        """
-        NAME = self.parent.NAME
-        self._parameter_mapping = get_recv_mapping(params, NAME, recv_all=True)
-        self._params = type(params).from_template(self._parameter_mapping.values())
-        self._params.update_kw_parameters(
-            {k: params.get_param(k) for k in self._params.keys()}
-        )
-        self.__recv_mapping = get_recv_mapping(params, NAME)
-        self.__send_mapping = get_send_mapping(params, NAME)
-
-    def get_new_inputs(self, remapper: Optional[Union[Callable, Dict]] = None):
-        """
-        Get new key mappings from the ParameterFrame.
-
-        Parameters
-        ----------
-        remapper: Optional[Union[callable, dict]]
-            a function or dictionary for remapping variable names.
+        remapper: Optional[Union[Callable, Dict[str, str]]]
+            A function or dictionary for remapping variable names.
             Useful for renaming old variables
 
         Returns
         -------
-        _inputs: dict
-            key value pairs of external program variable names and values
-
-        TODO unit conversion
+        _inputs: Dict[str, float]
+            Keys are external code parameter names, values are the input
+            values for those parameters.
         """
         _inputs = {}
 
         if not (callable(remapper) or isinstance(remapper, (type(None), Dict))):
             raise TypeError("remapper is not callable or a dictionary")
-        elif isinstance(remapper, Dict):
+        if isinstance(remapper, Dict):
             orig_remap = remapper.copy()
 
             def remapper(x):
@@ -177,278 +94,175 @@ class Setup(Task):
             def remapper(x):
                 return x
 
-        for prog_key, bm_key in self._send_mapping.items():
-            prog_key = remapper(prog_key)
-            if isinstance(prog_key, list):
-                for key in prog_key:
+        send_mappings = get_send_mapping(self.params, self._name)
+        for external_key, bm_key in send_mappings.items():
+            external_key = remapper(external_key)
+            if isinstance(external_key, list):
+                for key in external_key:
                     _inputs[key] = self._convert_units(self.params.get_param(bm_key))
                 continue
 
-            _inputs[prog_key] = self._convert_units(self.params.get_param(bm_key))
+            _inputs[external_key] = self._convert_units(self.params.get_param(bm_key))
 
         return _inputs
 
     def _convert_units(self, param):
-        code_unit = param.mapping[self.parent.NAME].unit
+        code_unit = param.mapping[self._name].unit
         if code_unit is not None:
             return raw_uc(param.value, param.unit, code_unit)
         else:
             return param.value
 
-    @property
-    def _recv_mapping(self):
-        self.__recv_mapping = get_recv_mapping(self.params, self.parent.NAME)
-        return self.__recv_mapping
 
-    @property
-    def _send_mapping(self):
-        self.__send_mapping = get_send_mapping(self.params, self.parent.NAME)
-        return self.__send_mapping
-
-    @property
-    def params(self) -> ParameterFrame:
-        """
-        The ParameterFrame corresponding to this run.
-        """
-        return self._params
-
-
-class Run(Task):
+class CodesTeardown(CodesTask):
     """
-    Generic Run Task
+    Base class for teardown tasks of a solver for an external code.
 
     Parameters
     ----------
-    parent: obj
-        parent file interface object
-    binary: str
-        binary location
+    params: ParameterFrame
+        The parameters for this task.
+    codes_name: str
+        The name of the external code the task is associated with.
     """
 
-    _binary = None
-
-    def __init__(self, parent, binary=None, *args, **kwargs):
-        super().__init__(parent)
-        if binary is not None:
-            self._binary = binary
-        if self._binary is None:
-            raise CodesError(f"Binary for {self.parent.NAME} not defined")
-
-
-class Teardown(Task):
-    """
-    Generic Teardown Task
-
-    Parameters
-    ----------
-    parent: obj
-        parent file interface object
-    """
-
-    def __init__(self, parent, *args, **kwargs):
-        super().__init__(parent)
-
-    def prepare_outputs(self, outputs: Dict, source: Optional[str] = None):
-        """
-        Prepare outputs for ParameterFrame
-
-        Implicitly converts to bluemira units in unit available
-
-        Parameters
-        ----------
-        outputs: Dict
-            key value pair of code outputs
-        source: Optional[str]
-            Set the source of all outputs, by default is code name
-
-        """
-        for bm_key, value in outputs.items():
-            try:
-                code_unit = (
-                    self.parent.params.get_param(bm_key).mapping[self.parent.NAME].unit
-                )
-            except AttributeError as exc:
-                raise AttributeError(f"No mapping found for {bm_key}") from exc
-
-            if value is None:
-                bluemira_warn(f"No value for {bm_key} found in output.")
-            else:
-                if code_unit is not None:
-                    outputs[bm_key] = {"value": value, "unit": code_unit}
-
-        self.parent.params.update_kw_parameters(
-            outputs, source=source if source is not None else f"{self.parent.NAME}"
-        )
-
-
-class FileProgramInterface(ABC):
-    """An external code wrapper"""
-
-    _setup = Setup
-    _run = Run
-    _teardown = Teardown
-    _runmode = RunMode
-
-    def __init__(
-        self,
-        NAME,
-        params,
-        runmode,
-        *args,
-        run_dir=None,
-        read_dir=None,
-        mappings=None,
-        problem_settings=None,
-        **kwargs,
+    def _update_params_with_outputs(
+        self, outputs: Dict[str, float], recv_all: bool = False
     ):
-        self.NAME = NAME
-
-        add_mapping(NAME, params, mappings)
-
-        if not hasattr(self, "run_dir"):
-            self.run_dir = "./" if run_dir is None else run_dir
-
-        self.read_dir = read_dir
-
-        if self._runmode is not RunMode and issubclass(self._runmode, RunMode):
-            self._set_runmode(runmode)
-        else:
-            raise CodesError("Please define a RunMode child class")
-
-        self._protect_tasks()
-
-        self.setup_obj = self._setup(
-            self, *args, params=params, problem_settings=problem_settings, **kwargs
-        )
-        self.run_obj = self._run(self, *args, **kwargs)
-        self.teardown_obj = self._teardown(self, *args, **kwargs)
-
-    def _protect_tasks(self):
         """
-        If tasks are not a child class then use the defaults.
-        """
-        for sub_name, parent in [
-            ["_setup", Setup],
-            ["_run", Run],
-            ["_teardown", Teardown],
-        ]:
-            if not issubclass(getattr(self, sub_name), parent):
-                bluemira_warn("Using default {parent.__name__} task")
-                setattr(self, sub_name, parent)
+        Update this task's parameters with the external code's outputs.
 
-    @property
-    def binary(self):
-        """
-        Program binary name
-        """
-        return self.run_obj._binary
-
-    @binary.setter
-    def binary(self, _binary: str):
-        """
-        Set program binary name
+        This implicitly performs any unit conversions.
 
         Parameters
         ----------
-        _binary: str
-            binary name
+        outputs: Dict[str, float]
+            Key are the external code's parameter names, the values are
+            the values for those parameters.
+        recv_all: bool
+            Whether to ignore the 'recv' attribute on the parameter
+            mapping, and update all output parameter values.
 
+        Raises
+        ------
+        CodesError
+            If any output does not have a mapping to a bluemira
+            parameter, or the output maps to a bluemira parameter that
+            does not exist in this object's ParameterFrame.
         """
-        self.run_obj._binary = _binary
+        mapped_outputs = self._map_external_outputs_to_bluemira_params(outputs, recv_all)
+        self.params.update_kw_parameters(mapped_outputs, source=self._name)
 
-    def _set_runmode(self, runmode: str):
+    def _map_external_outputs_to_bluemira_params(
+        self, external_outputs: Dict[str, Any], recv_all: bool
+    ) -> Dict[str, Dict[str, Any]]:
         """
-        Set the runmode
+        Loop through external outputs, find the corresponding bluemira
+        parameter name, and map it to the output's value and unit.
 
         Parameters
         ----------
-        runmode: str
-            runmode to be set
-        """
-        mode = runmode.upper().translate(str.maketrans("", "", string.whitespace))
-        self._runner = self._runmode[mode]
+        external_outputs: Dict[str, Any]
+            An output produced by an external code. The keys are the
+            outputs' names (not the bluemira version of the name), the
+            values are the output's value (in the external code's unit).
+        recv_all: bool
+            Whether to ignore the 'recv' attribute on the parameter
+            mapping, and update all output parameter values.
 
-    @property
-    def params(self) -> ParameterFrame:
+        Returns
+        -------
+        mapped_outputs: Dict[str, Dict[str, Any]]
+            The keys are bluemira parameter names and the values are a
+            dict of form '{"value": Any, "unit": str}', where the value
+            is the external code's output value, and the unit is the
+            external code's unit.
         """
-        The ParameterFrame corresponding to this run.
-        """
-        return self.setup_obj.params
+        mapped_outputs = {}
+        recv_mappings = get_recv_mapping(self.params, self._name, recv_all)
+        for external_key, bluemira_key in recv_mappings.items():
+            output_value = self._get_output_or_raise(external_outputs, external_key)
+            if output_value is None:
+                continue
+            param_mapping = self._get_parameter_mapping_or_raise(bluemira_key)
+            if param_mapping.unit is not None:
+                mapped_outputs[bluemira_key] = {
+                    "value": output_value,
+                    "unit": param_mapping.unit,
+                }
+        return mapped_outputs
 
-    @property
-    def _parameter_mapping(self) -> Dict[str, str]:
-        """
-        The ParameterFrame corresponding to this run.
-        """
-        return self.setup_obj._parameter_mapping
+    def _get_output_or_raise(
+        self, external_outputs: Dict[str, Any], parameter_name: str
+    ):
+        try:
+            output_value = external_outputs[parameter_name]
+        except KeyError as key_error:
+            raise CodesError(
+                f"No output value from code '{self._name}' found for parameter "
+                f"'{parameter_name}'."
+            ) from key_error
+        if output_value is None:
+            bluemira_warn(
+                f"No value for output parameter '{parameter_name}' from code "
+                f"'{self._name}'."
+            )
+        return output_value
 
-    @property
-    def _recv_mapping(self) -> Dict[str, str]:
-        """
-        The ParameterFrame corresponding to this run.
-        """
-        return self.setup_obj._recv_mapping
+    def _get_parameter_mapping_or_raise(self, bluemira_param_name: str):
+        try:
+            return self.params.get_param(bluemira_param_name).mapping[self._name]
+        except AttributeError as attr_error:
+            raise CodesError(
+                f"No mapping defined between parameter '{bluemira_param_name}' and "
+                f"code '{self._name}'."
+            ) from attr_error
 
-    @property
-    def _send_mapping(self) -> Dict[str, str]:
-        """
-        The ParameterFrame corresponding to this run.
-        """
-        return self.setup_obj._send_mapping
 
-    @property
-    def problem_settings(self):
-        """
-        Get problem settings dictionary
-        """
-        return self.setup_obj._problem_settings
+class CodesSolver(SolverABC):
+    """
+    Base class for solvers running an external code.
+    """
 
-    def modify_mappings(self, mappings: Dict[str, Dict[str, bool]]):
+    @abc.abstractproperty
+    def name(self):
         """
-        Modify the send/recieve mappings of a key
+        The name of the solver.
+
+        In the base class, this is used to find mappings and specialise
+        error messages for the concrete solver.
+        """
+        pass
+
+    def modify_mappings(self, send_recv: Dict[str, Dict[str, bool]]):
+        """
+        Modify the send/receive truth values of a parameter.
+
+        If a parameter's 'send' is set to False, its value will not be
+        passed to the external code (a default will be used). Likewise,
+        if a parameter's 'recv' is False, its value will not be updated
+        from the external code's outputs.
 
         Parameters
         ----------
         mappings: dict
-            A dictionary of variables to change mappings.
+            A dictionary where keys are variables to change the mappings
+            of, and values specify 'send', and or, 'recv' booleans.
 
-        Notes
-        -----
-            Only one of send or recv is needed. The mappings dictionary could look like:
+            E.g.,
 
-               {"var1": {"send": False, "recv": True}, "var2": {"recv": False}}
+            .. code-block:: python
 
+                {
+                    "var1": {"send": False, "recv": True},
+                    "var2": {"recv": False}
+                }
         """
-        for key, val in mappings.items():
+        for key, val in send_recv.items():
             try:
-                p_map = getattr(self.params, key).mapping[self.NAME]
+                p_map = getattr(self.params, key).mapping[self.name]
             except (AttributeError, KeyError):
-                bluemira_warn(f"No mapping known for {key} in {self.NAME}")
+                bluemira_warn(f"No mapping known for {key} in {self.name}")
             else:
                 for sr_key, sr_val in val.items():
                     setattr(p_map, sr_key, sr_val)
-
-    def run(self, *args, **kwargs):
-        """
-        Run the full program interface
-        """
-        self._runner(self.setup_obj, *args, **kwargs)
-        self._runner(self.run_obj, *args, **kwargs)
-        self._runner(self.teardown_obj, *args, **kwargs)
-
-    @abstractmethod
-    def get_raw_variables(self, params: Union[List, str]):
-        """
-        Get raw parameters from an external code
-        (mapped bluemira parameters will have bluemira names)
-
-        Parameters
-        ----------
-        params: Union[List, str]
-            parameter names to access
-
-        Returns
-        -------
-        values list
-        """
-        pass
