+++ bluemira/codes/plasmod/api/_run.py
@@ -25,7 +25,7 @@ Defines the 'Run' stage of the plasmod solver.
 from bluemira.base.look_and_feel import bluemira_print
 from bluemira.base.parameter import ParameterFrame
 from bluemira.codes.error import CodesError
-from bluemira.codes.interface_ import CodesTask
+from bluemira.codes.interface import CodesTask
 from bluemira.codes.plasmod.constants import BINARY as PLASMOD_BINARY
 from bluemira.codes.plasmod.constants import NAME as PLASMOD_NAME
