+++ bluemira/codes/process/_run.py
@@ -25,7 +25,7 @@ PROCESS run functions
 
 from bluemira.base.look_and_feel import bluemira_print
 from bluemira.base.parameter import ParameterFrame
-from bluemira.codes.interface_ import CodesTask
+from bluemira.codes.interface import CodesTask
 from bluemira.codes.process.constants import BINARY as PROCESS_BINARY
 from bluemira.codes.process.constants import NAME as PROCESS_NAME
